+++ apps/client/src/app/spec-download/spec-download.component.ts
@@ -18,12 +18,10 @@ interface PayloadDownload {
 })
 export class SpecDownloadComponent implements OnInit {
   request: any = {}
-  loadingDownloadFiles: boolean = false;
+  loadingDownloadJSON: boolean = false;
+  loadingDownloadXML: boolean = false;
   loadingDownloadSpreadsheet: boolean = false;
   version = 'ecrv3'
-  bundleType = ''
-  contentType = 'json'
-  isDisabled = true;
 
   constructor(
     private httpClient: HttpClient,
@@ -34,78 +32,40 @@ export class SpecDownloadComponent implements OnInit {
   ngOnInit() {
   }
 
-  setVersion(e) {
-    this.version = e.target.value
-  } // eRSD (eCR) V1 or V2
-
-  setBundle(e) {
-    this.bundleType = e.target.value
-  } // Supplemental or Specification
-
-  // setContentType(e) { this.contentType = e.target.value } // XML and/or JSON
-
-  //   setContentType(e) {
-  //     const value = e.target.value;
-  //     const isChecked = e.target.checked;
-
-  //     // Check if the checkbox is checked
-  //     if (isChecked) {
-  //         // Add the value to the contentType array (if not already present)
-  //         if (!this.contentType.includes(value)) {
-  //             this.contentType.push(value); 
-  //         }
-  //     } else {
-  //         // Remove the value from the contentType array (if present)
-  //         const index = this.contentType.indexOf(value);
-  //         if (index !== -1) {
-  //             this.contentType.splice(index, 1);
-  //         }
-  //     }
-  //     // Check if no checkboxes are checked and set isDisabled accordingly
-  //     this.isDisabled = this.contentType.length === 0;
-  // }
-
-  setContentType(e) {
-    const value = e.target.value;
-    // Set the selected content type
-    this.contentType = value;
-    // Check if no radio button is selected and set isDisabled accordingly
-    // this.isDisabled = this.contentType === null;
-  }
-
-  // buildFileName() {
-  //   if (this.bundleType !== '') {
-  //     return `${this.version}-${this.bundleType}.${this.contentType}`
-  //   }
-  //   return `${this.version}.${this.contentType}`
-  // }
-
-  async handleSubmit() {
-    let url = '';
+  async downloadJSON() {
     try {
-      if (this.version) {
-        // when V2 Supplemental goes live: if(this.bundleType == '') { throw Error('Please select a bundle type') } 
-        url = `/api/s3/${this.contentType}?version=${this.version}`
-        // when V2 Supplemental goes live: url = `/api/s3/${this.contentType}?version=${this.version}&bundle=${this.bundleType}`
-      }
-      // else {
-      //   url = `/api/s3/${this.contentType}?version=${this.version}`;
-      // }
-      // return this.queryServer(url);
-      this.loadingDownloadFiles = true;
+      const url = `/api/s3/json?version=${this.version}`;
+      this.loadingDownloadJSON = true;
       const data = await firstValueFrom(this.httpClient.post(url, this.request)) as PayloadDownload;
-      this.loadingDownloadFiles = false;
+      this.loadingDownloadJSON = false;
       await this.downloadS3(data);
-
     } catch (err) {
-      alert(`Error while downloading file: ${err.message}`);
+      this.loadingDownloadJSON = false;
+      alert(`Error while downloading JSON file: ${err.message}`);
       console.error(err);
     }
   }
 
-  async downloadReleaseNotes(description) {
+  async downloadXML() {
     try {
-      let url = `api/download/release_notes?version=${description}`;
+      const url = `/api/s3/xml?version=${this.version}`;
+      this.loadingDownloadXML = true;
+      const data = await firstValueFrom(this.httpClient.post(url, this.request)) as PayloadDownload;
+      this.loadingDownloadXML = false;
+      await this.downloadS3(data);
+    } catch (err) {
+      this.loadingDownloadXML = false;
+      alert(`Error while downloading XML file: ${err.message}`);
+      console.error(err);
+    }
+  }
+
+  async downloadReleaseNotes(description: string) {
+    try {
+      if (description !== 'ersdv3') {
+        throw new Error('Only eRSD Version 3 release notes are available');
+      }
+      const url = `api/download/release_notes?version=${description}`;
       const data = await firstValueFrom(this.httpClient.post(url, this.request)) as PayloadDownload;
       await this.downloadS3(data);